added mdr, begin of naming cols like fhir resources, csv import, dashboard

+++ DQA_Tool/app/server.R
@@ -1,43 +1,55 @@
 shinyServer(function(input, output, session) {
-
+    
     # if you want to source any files with functions, do it inside the server-function, so the information will not be shared across sessions
     source("./_utilities/functions.R", encoding = "UTF-8")
     source("./_utilities/statistics.R", encoding = "UTF-8")
     
     # define reactive values here
     rv <- reactiveValues(
+        mdr = NULL,
         file = NULL,
         sourcefiledir = NULL,
         sitename = NULL,
         db_settings = NULL,
         db_con = NULL,
-        db_getdata = NULL,
+        target_getdata = NULL,
+        source_getdata = NULL,
         data_objects = list(),
-        dash_summary_target = data.table("variable" = character(), 
-                                  "distinct" = integer(), 
-                                  "valids" = integer(),
-                                  "missings" = integer())
+        dash_summary_target = summaryTable(),
+        dash_summary_source = summaryTable(),
+        csv_keys = c("FALL.CSV", "FAB.CSV", "ICD.CSV", "OPS.CSV")
     )
     
+    
     # run onStart here
-    onStart(session, rv, output)
+    onStart(session, rv, input, output)
+    
+    # read mdr
+    observe({
+        if (is.null(rv$mdr)){
+            cat("\nRead MDR\n")
+            rv$mdr <- fread("./_utilities/DQ_MDR.csv", header = T)
+        }
+    })
+    
     
     # handle reset
     observeEvent(input$reset, {
         js$reset()
     })
-
+    
+    
     
     ########################
     # tab_config
     ########################
     # observe source file directory
-    observe({
-        if (identical(rv$sourcefiledir, character(0)) || is.null(rv$sourcefiledir)){
-            shinyDirChoose(input, "config_sourcedir_in", updateFreq = 0, session = session, defaultPath = "", roots = c(home="/home/"), defaultRoot = "home")
-            roots = c(home="/home/")
-            rv$sourcefiledir <- parseDirPath(roots, input$config_sourcedir_in)
-        }
+    observeEvent(input$config_sourcedir_in, {
+        shinyDirChoose(input, "config_sourcedir_in", updateFreq = 0, session = session, defaultPath = "", roots = c(home="/home/"), defaultRoot = "home")
+        #if (identical(rv$sourcefiledir, character(0)) || is.null(rv$sourcefiledir)){
+        roots = c(home="/home/")
+        rv$sourcefiledir <- parseDirPath(roots, input$config_sourcedir_in)
+        #}
     })
     
     output$config_sourcedir_out <- reactive({
@@ -51,8 +63,11 @@ shinyServer(function(input, output, session) {
         
         # remove existing global_settings here as they will newly be created once loading the data
         if (file.exists("_settings/global_settings.JSON")){
-            cat("\nRemoving '_settings/global_settings.JSON'")
-            file.remove("_settings/global_settings.JSON")
+            if (input$config_targetdb_rad != rv$user_settings[["db"]]){
+                cat("\nRemoving '_settings/global_settings.JSON'")
+                file.remove("_settings/global_settings.JSON")
+                rv$user_settings <- NULL
+            }
         }
         
         # if "./_utilities/settings.yml" not present, read default settings list here and populate textInputs
@@ -61,7 +76,7 @@ shinyServer(function(input, output, session) {
             rv$db_settings <- config::get(input$config_targetdb_rad, file = "./_utilities/settings_default.yml")
             
             showModal(modalDialog(
-                "Loading default configuration", 
+                "Loading default configuration.", 
                 title = "Loading default database configuration")
             )
         } else {
@@ -78,7 +93,7 @@ shinyServer(function(input, output, session) {
     # observe saving of settings
     observeEvent(input$config_targetdb_save_btn,{
         rv$db_settings <- getDBsettings(input, rv)
-            
+        
         if (!is.null(rv$db_settings)){
             print(rv$db_settings)
             
@@ -100,7 +115,7 @@ shinyServer(function(input, output, session) {
         if (!is.null(rv$db_settings)){
             
             testDBcon(rv)
-        
+            
             if (!is.null(rv$db_con)){
                 cat("\nDB connection successfully established\n")
                 showModal(modalDialog(
@@ -154,7 +169,7 @@ shinyServer(function(input, output, session) {
         if (nchar(input$config_sitename) < 2 || any(grepl("\\s", input$config_sitename))){
             showModal(modalDialog(
                 title = "Invalid values",
-                "No empty strings or spaces allowed in the site name configuration"
+                "No empty strings or spaces allowed in the site name configuration."
             ))
             rv$sql <- NULL
         } else {
@@ -165,20 +180,40 @@ shinyServer(function(input, output, session) {
             if (identical(rv$sourcefiledir, character(0)) || any(grepl("\\s", rv$sourcefiledir))){
                 showModal(modalDialog(
                     title = "Invalid values",
-                    "No empty strings or spaces allowed in the source file path configuration"
+                    "No empty strings or spaces allowed in the source file path configuration."
                 ))
                 rv$sql <- NULL
             } else {
-                # TODO check for required files here: FALL.CSV, FAB.CSV, ICD.CSV, OPS.CSV
-                cat("\nPrint list of files in sourcefiledir:\n")
                 filelist <- list.files(path=rv$sourcefiledir, pattern = "\\.CSV|\\.csv", full.names = T)
-                print(filelist)
+                # iterate over list and check for presence of required filenames: FALL.CSV, FAB.CSV, ICD.CSV, OPS.CSV
+                check <- sapply(filelist, function(i){
+                    cat("\n", i, "\n")
+                    return(grepl("FALL\\.CSV$|FAB\\.CSV$|ICD\\.CSV$|OPS.CSV$", i))
+                })
+                
+                tryCatch({
+                    # test if there are exact 4 source files
+                    if (sum(check)!=4){
+                        showModal(modalDialog(
+                            title = "Invalid path",
+                            "The specified directory does not contain the 4 neccessary CSV-files (FALL.CSV, FAB.CSV, ICD.CSV, OPS.CSV)."
+                        ))  
+                        rv$sql <- NULL
+                    } 
+                }, error = function(e){
+                    showModal(modalDialog(
+                        title = "Invalid path",
+                        "There are no CSV-files in the specified directory."
+                    ))
+                    rv$sql <- NULL
+                })
             }
         }
         
         
         if (!is.null(rv$sql)){
-            rv$db_getdata <- TRUE
+            rv$target_getdata <- TRUE
+            rv$source_getdata <- TRUE
             
             # disable config page
             shinyjs::disable("config_targetdb_rad")
@@ -221,27 +256,68 @@ shinyServer(function(input, output, session) {
         }
     })
     
-    # calculate number of distinct patient ids
+    # calculate overview
+    # target
     observe({
-        req(rv$dt_patient.db)
-        if (!("patient_id" %in% rv$dash_summary_target[,variable])){
-            rv[["dt_patient.db_summary"]] <- countUnique(rv$dt_patient.db, "patient_id", input$config_targetdb_rad)
-            rv$dash_summary_target <- rbind(rv$dash_summary_target, rv$dt_patient.db_summary[,.(variable, distinct, valids, missings)])
+        req(rv$dt.patient_target)
+        if (nrow(rv$dash_summary_target) < 2){
+            cat("\nBuild rv$dash_summary_target")
+            if (!("patient.identifier.value" %in% rv$dash_summary_target[,variable])){
+                rv[["ov.patient_target.summary"]] <- countUnique(rv$dt.patient_target, "patient_identifier_value", input$config_targetdb_rad)
+                rv$dash_summary_target <- rbind(rv$dash_summary_target, rv$ov.patient_target.summary[,.(variable, distinct, valids, missings)])
+            }
         }
     })
     observe({
-        req(rv$dt_visit.db)
-        if (!("encounter_id" %in% rv$dash_summary_target[,variable])){
-            rv[["dt_visit.db_summary"]] <- countUnique(rv$dt_visit.db, "encounter_id", input$config_targetdb_rad)
-            rv$dash_summary_target <- rbind(rv$dash_summary_target, rv$dt_visit.db_summary[,.(variable, distinct, valids, missings)])
+        req(rv$dt.encounter_target)
+        if (nrow(rv$dash_summary_target) < 2){
+            cat("\nBuild rv$dash_summary_target2")
+            if (!("encounter.identifier.value" %in% rv$dash_summary_target[,variable])){
+                rv[["ov.encounter_target.summary"]] <- countUnique(rv$dt.encounter_target, "encounter_identifier_value", input$config_targetdb_rad)
+                rv$dash_summary_target <- rbind(rv$dash_summary_target, rv$ov.encounter_target.summary[,.(variable, distinct, valids, missings)])
+            }
         }
     })
-
+    
+    # source
+    observe({
+        req(rv$FALL.CSV)
+        if (nrow(rv$dash_summary_source) < 2){
+            cat("\nBuild rv$dash_summary_source")
+            if (!("PATIENTENNUMMER" %in% rv$dash_summary_source[,variable])){
+                rv[["ov.patient_source.summary"]] <- countUnique(rv$FALL.CSV, "PATIENTENNUMMER", "csv")
+                rv$dash_summary_source <- rbind(rv$dash_summary_source, rv$ov.patient_source.summary[,.(variable, distinct, valids, missings)])
+            }
+            if (!("KH_INTERNES_KENNZEICHEN" %in% rv$dash_summary_source[,variable])){
+                rv[["ov.encounter_source.summary"]] <- countUnique(rv$FALL.CSV, "KH_INTERNES_KENNZEICHEN", "csv")
+                rv$dash_summary_source <- rbind(rv$dash_summary_source, rv$ov.encounter_source.summary[,.(variable, distinct, valids, missings)])
+            }
+            if (!("Begleitpersonen" %in% rv$dash_summary_source[,variable])){
+                tab <- countUnique(rv$FALL.CSV[AUFNAHMEANLASS=="B",], "KH_INTERNES_KENNZEICHEN", "csv")
+                if (nrow(tab) == 0){
+                    cat("\nThere are no chaperones present in your data.\n")
+                } 
+                rv[["ov.chaperone_source.summary"]] <- tab[1,variable:="Begleitpersonen"]
+                rv$dash_summary_source <- rbind(rv$dash_summary_source, rv$ov.chaperone_source.summary[,.(variable, distinct, valids, missings)])
+            }
+        }
+    })
+    
+    
+    
     # render dashboard summary
     observe({
         if(nrow(rv$dash_summary_target) > 0) {
             output$dash_summary_target <- renderTable({
                 tab <- rv$dash_summary_target
+                colnames(tab) <- c("Variable", "# Distinct", "# Valid", "# Missing")
+                tab
+            })
+        }
+        
+        if(nrow(rv$dash_summary_source) > 0) {
+            output$dash_summary_source <- renderTable({
+                tab <- rv$dash_summary_source
                 colnames(tab) <- c("Variable", "# Distinct", "# Valid", "# Missing")
                 tab
             })
@@ -252,24 +328,46 @@ shinyServer(function(input, output, session) {
     # tab_rawdata1
     ########################
     observe({
-        req(rv$db_getdata)
-        vec <- c("dt_patient.db", "dt_visit.db", "dt_aufnan.db") #, 
-                 #"dt_aitaa.db", "dt_aijaa.db", "dt_aufngr.db", "dt_entlgr.db", "dt_beatmst.db") #,
-                 # "dt_icd.db", "dt_ops.db", "dt_fab.db",
-                 # "dt_pl_c5x.db", "dt_pl_c6x.db", "dt_pl_05xx.db", "dt_pl_o0099.db")
+        req(rv$target_getdata)
+        # get data from database
         
-        for (i in vec){
+        # get keys from our mdr
+        rv$db_keys <- rv$mdr[source_system==input$config_targetdb_rad,unique(key)][1:3]
+        
+        for (i in rv$db_keys){
             if (is.null(eval(parse(text=paste0("rv$", i))))){
                 rv[[i]] <- fireSQL(rv, i)
             }
         }
         invisible(gc())
-        rv$db_getdata <- FALSE
+        rv$target_getdata <- FALSE
+    })
+    
+    observe({
+        req(rv$source_getdata)
+        # get data from csv files
+        cat("\nReading from csv\n")
+        
+        for (i in rv$csv_keys){
+            if (is.null(eval(parse(text=paste0("rv$", i))))){
+                rv[[i]] <- loadCSV(rv, i)
+            }
+        }
+        
+        invisible(gc())
+        rv$source_getdata <- FALSE
     })
     
     observeEvent(input$rawdata1_sel, {
         output$rawdata1_table <- DT::renderDataTable({
             DT::datatable(eval(parse(text= input$rawdata1_sel)), options = list(scrollX = TRUE, pageLength = 20))
         })
+    })
+    
+    ########################
+    # tab_mdr
+    ########################
+    output$mdr_table <- DT::renderDataTable({
+        DT::datatable(rv$mdr, options = list(scrollX = TRUE, pageLength = 20))
     })
 })
